Add files via upload
1. The web application must use the colour scheme of the foundation – purple and orange. 
2. Users must log in securely to the web application to be able to edit information. 
3. The web application must allow authorised users to capture new monetary donations. The 
date and amount are mandatory, but the donor may decide to remain anonymous. 
4. The web application must allow authorised users to capture new goods donations. The 
date, number of items, category, and description of each item are mandatory. But the 
donor may decide to remain anonymous.
The web application must be pre-configured with the following categories of goods: 
a. Clothes 
b. Non-perishable foods 
6. The web application must allow authorised users to define new categories of goods. 
7. The web application must allow authorised users to capture a new disaster. The data for a 
disaster includes the start date and end date, the location, and a description. It must also 
be possible to specify the required aid types, for example, water provision, clothing, food, 
etc. 
8. The web application must allow authorised users to view the following lists: 
a. All incoming monetary donations. 
b. All incoming goods donations. 
c. All disasters.